+++ contract.py
@@ -29,21 +29,6 @@ def approval_program():
             ),
         )
 
-    @Subroutine(TealType.none)
-    def close_account_to_creator() -> Expr:
-        return If(Balance(Global.current_application_address()) != Int(0)).Then(
-            Seq(
-                InnerTxnBuilder.Begin(),
-                InnerTxnBuilder.SetFields(
-                    {
-                        TxnField.type_enum: TxnType.Payment,
-                        TxnField.close_remainder_to: Global.creator_address(),
-                    }
-                ),
-                InnerTxnBuilder.Submit(),
-            )
-        )
-
     on_create = Seq(
         Approve(),
     )
@@ -52,8 +37,6 @@ def approval_program():
         App.globalPut(seller_address_key, Txn.application_args[1]),
         App.globalPut(nft_id_key, Btoi(Txn.application_args[2])),
         App.globalPut(price_key, Btoi(Txn.application_args[3])),
-        # Assert(Global.creator_address() == Txn.sender),
-        # Assert(Global.latest_timestamp() < App.globalGet(start_time_key)),
         InnerTxnBuilder.Begin(),
         InnerTxnBuilder.SetFields(
             {
@@ -68,7 +51,6 @@ def approval_program():
 
     on_buy = Seq(
         App.globalPut(buyer_address_key, Txn.application_args[1]),
-        # Assert(Global.creator_address() == Txn.sender),
         close_nft_to(
             App.globalGet(nft_id_key),
             App.globalGet(buyer_address_key)
@@ -84,7 +66,6 @@ def approval_program():
         InnerTxnBuilder.Submit(),
         Approve(),
     )
-
 
     on_call = Cond(
         [Txn.application_args[0] == Bytes("setup"), on_setup],
@@ -109,12 +90,3 @@ def approval_program():
 
 def clear_state_program():
     return Approve()
-
-# if __name__ == "__main__":
-#     with open("auction_approval.teal", "w") as f:
-#         compiled = compileTeal(approval_program(), mode=Mode.Application, version=5)
-#         f.write(compiled)
-
-#     with open("auction_clear_state.teal", "w") as f:
-#         compiled = compileTeal(clear_state_program(), mode=Mode.Application, version=5)
-#         f.write(compiled)
